baseline/hyoentae-experiment/requirements.txt


# --- Core numeric stack (NumPy 2 계열로 정렬) ---
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2

# --- CV / Augmentations ---
opencv-python-headless==4.12.0.88   # 4.8.1.78 이상 권장
albumentations==1.4.16
augraphy==8.2.6
Pillow>=10.0.0

# --- DL / Vision ---
torch==2.9.0
torchvision==0.24.0
timm==0.9.12

# --- Experiment management ---
hydra-core==1.3.2
wandb>=0.17.0

# --- Jupyter / plotting (선호 버전) ---
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.1
jupyter==1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# --- Utils ---
tqdm
PyYAML>=6.0
